fixed the main()
Also, i think i changed some other things but i dont remember. And i changed how the results are print because i wanted to make a tabulate object, but i failed, so i made it create a string for every result and print it. That way i think we can manipulate it easier than before.